Separate model inference, add train/test data
Each of the sub-models in the Bayesian network is now separately estimated by pymc3, which greatly improved the speed of inference compared to the previous approach. Upon completion of inference, model information is stored in pickled files which can then be reloaded later. Upon reloading, each model is reinstantiated with new (out of sample) input data that is used for testing predictions.

A new /data folder now includes the synthetic data from EnergyPlus that are used to train each of the Bayesian sub-models, as well as a set of data to use in testing predicted control choices under different sets of input conditions.