Update prediction data, add null option, rework updating routine
The file “test_predict.csv” was updated to correct the “h_DR_Start” variable (first hour of event is now 1, second is 2, etc.), carry through incentives to the rebound hours (so that any increase in energy during rebound is penalized from a cost perspective), and limit the rebound period to just 1 hour. Additionally, the “Baseline - Nothing” option was removed from this file.

Instead of representing the baseline (null) option in the “test_predict.csv” input file, this option is now handled in flex.py as a measure that always yields zero change in the input variables to the operator choice function. It is still written out to “recommendations.json” as the measure name “Baseline - Nothing”.

Additionally, the model updating routine was reworked for ease of use in the cosimulation testing. Most of the code for this is moved into a function “gen_updates” that yields updated model parameter traces for each model type (currently we update the demand and temperature models only), and accepts as input the starting and ending event number in the “test_update.csv” file (see column “Day.number”) from which to pull data for the update. After the last update is finished (e.g., when the ending event equals the total expected number of events), diagnostic plots for each model type are generated and written to the /.data folder (e.g., ./diagnostic_plots/update_dmd.png and ./diagnostic_plots/update_tmp.png”).