Adjust handling of multi-hour predictions, remove null option, add diagnostic plots
The operator choice function is now only called once in the prediction routine (as opposed to during each hour reflected in test_predict.csv), and it uses the following inputs predicted across the hours reflected in the “test_predict.csv file:

* Total economic benefit (100$)
* Maximum temperature change during event (deg F)
* Maximum pre-cool temperature change (deg F)
* Maximum lighting change during event (frac)
* Outdoor air ventilation fraction change during event (currently zero across event hours)
* Plug load power change during event (constant across event hours)

The null option was removed entirely from the simulation, as it was determined that this option is inconsistent with how we’ve fit the demand/service models. Its selection is implied when the model cannot decide across the other (non-null) options and the operator’s default preference is to do nothing.

The example incentive level during event hours is now $0.25/kWh and $0.1/kWh in the rebound hour in test_predict.csv.

An R script that generate a set of boxplots for input/output values in the ‘recommendations.json’ file generated by flex.py was added.