Add baseline demand models
Models that can predict baseline hourly demand (W/sf) for medium office and retail are added and incorporated into the model initialization, assessment, updating, and prediction routines.

The latter is executed by adding a new ‘--base_pred’ option at the command line in the absence of the options for model initialization, assessment, or updating.

The previous ‘--no_base’ option was changed to ‘--null_strategy’ to avoid confusion, and the default is now to not include this null strategy in the recommendation routine (where before it was included by default).