Add screening of candidate strategies, return to pre-cooling choice bins
When a candidate set of strategies is entered into the prediction routine, the code will now screen out any strategies for which input conditions are invalid (outside the scope of the models). Currently, this means temperature/precooling strategies for an event in which the OAT goes below 70 degrees F will be screened out of consideration. In the output file, “recommendations.json” these strategies will still show up but with a choice probability of zero.

Additionally, the DCE model was re-estimated to once again reflect a binning of the pre-cooling coefficient for events in which the expected temperature raise is above or below 2 degrees F.